+++ src/components/datagrid/column_selector.tsx
@@ -27,13 +27,13 @@ import {
 } from '../drag_and_drop';
 import { DropResult } from 'react-beautiful-dnd';
 import { EuiIcon } from '../icon';
-import { useResettingState } from '../../services';
+import { useDependentState } from '../../services';
 
 export const useColumnSelector = (
   availableColumns: EuiDataGridColumn[],
   columnVisibility: EuiDataGridColumnVisibility
 ): [ReactElement, EuiDataGridColumn[]] => {
-  const [sortedColumns, setSortedColumns] = useResettingState(
+  const [sortedColumns, setSortedColumns] = useDependentState(
     () => availableColumns.map(({ id }) => id),
     [availableColumns]
   );