+++ src/services/hooks/index.ts
@@ -1,1 +1,1 @@
-export * from './useResettingState';
+export * from './useDependentState';